requirements.txt
pyperclip==1.8.2
python-dotenv==1.0.0
